@@ -23,25 +23,25 @@
     "ipfs-pubsub-room": "^3.0.0",
     "pako": "^2.0.4",
     "peerjs": "^1.4.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "valtio": "^1.4.0",
     "valtio-yjs": "^0.3.1",
     "wgo": "^3.0.0-alpha.10",
     "yjs": "^13.5.30"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/dompurify": "^2.3.3",
     "@types/emoji-mart": "^3.0.9",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.23",
-    "@types/pako": "^1.0.3",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
-    "@types/w3c-image-capture": "^1.0.5",
+    "@types/jest": "^29.0.0",
+    "@types/node": "^18.7.15",
+    "@types/pako": "^2.0.0",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "@types/w3c-image-capture": "^1.0.6",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
@@ -52,7 +52,7 @@
     "process": "^0.11.10",
     "react-scripts": "^5.0.1",
     "stream-browserify": "^3.0.0",
-    "typescript": "^4.6.3"
+    "typescript": "^4.8.2"
   },
   "eslintConfig": {
     "extends": [
